keyboard: move LocaledWrapper into localization module
